feat(inspector): own element-context capture; retire DevFeedbackWidget picker
ADR-007: the inspect-widget is the single element-selection + context
detector. Feedback surfaces consume what it emits instead of maintaining
parallel pickers.

inspect-widget.js
- Port page + section detection from the (now-deleted) detectContext.ts:
  the mockup `section--*` convention plus product-app semantic-landmark /
  aria / heading / id fallbacks. Component reuses the inspector's existing
  findBdsRoot — one component detector.
- Resolve { page, section, component, element_tag } for the selected element
  and dispatch it on `brik:inspect:report` (CustomEvent) for a host to consume.
- Add a "Feedback" action to the locked Click panel that fires the event.

DevFeedbackWidget.tsx
- Revert #879's pick-element UI (picking state, hover-highlight overlay,
  capture-on-click effect) — the widget is a pure submission form again.
- Keep the page/section/component payload (feedback-contract 0.2.0); expose
  them as props so a host can pre-fill from the inspector event. `page` still
  defaults to document.title.
- Delete detectContext.ts (was internal to the widget; not exported).

Refs brik-bds#880, brikdesigns/brik-llm#979
